#55 refactor: refactor video components
