+++ src/ControlComponents/ValueBox.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import Title from './Title.js';
+import Title from './Title.jsx';
 
 export default function ValueBox(props) {
   return (
