+++ src/VideoComponents/css/BiasWidget.css
@@ -1,0 +1,4 @@
+.BiasWidget {
+  right: 50px;
+  bottom: 50px;
+}
