+++ src/VideoComponents/css/CanvasWidget.css
@@ -1,0 +1,8 @@
+canvas {
+  position: fixed;
+}
+
+img {
+  position: fixed;
+  opacity: 0.5;
+}
