+++ src/VideoComponents/css/DepthWidget.css
@@ -1,0 +1,4 @@
+.DepthWidget {
+  top: 75px;
+  left: 10px;
+}
