--- src/VideoComponents/css/overlay.css
@@ -1,25 +1,0 @@
-canvas {
-  position: fixed;
-}
-
-img {
-  position: fixed;
-  opacity: 0.5;
-}
-
-#canvasBias {
-  right: 50px;
-  bottom: 50px;
-}
-
-#canvasHeading {
-  /* This centers the widget, even in the case where the windows becomes too small when resizing */
-  left: -9999px;
-  right: -9999px;
-  margin: auto;
-}
-
-#canvasDepth {
-  top: 75px;
-  left: 10px;
-}
